ref: Move messages folder into gui folder
