+++ ankihub/gui/deck_installation.py
@@ -16,11 +16,11 @@ from ..ankihub_client import NoteInfo
 from ..ankihub_client.ankihub_client import AnkiHubHTTPError
 from ..importing import AnkiHubImporter, AnkiHubImportResult
 from ..media_sync import media_sync
-from ..messages import messages
 from ..settings import config, url_view_deck
 from ..subdecks import deck_contains_subdeck_tags
 from ..utils import create_backup
 from .exceptions import DeckDownloadAndInstallError
+from .messages import messages
 from .subdecks import confirm_and_toggle_subdecks
 from .utils import ask_user
