+++ ankihub/gui/messages/templates/deck_install_confirmation.html
